refactor advisory workspace routes
