+++ src/app/routers/advisory_workspace_common.py
@@ -1,0 +1,7 @@
+from fastapi import Path
+
+WORKSPACE_ID_PATH = Path(
+    ...,
+    description="Advisory workspace identifier returned by lotus-advise.",
+    examples=["aws_001"],
+)
